Massive ovarhaul of CosPlanForm:
Create new component which took the original CosPlan component and built on the logic: now the form is structured to send data that the backend is able to receive it. Explicitly passing the formattedDueDate to the post request to ensure that Django is able to receive the expected date format (despite all of my settings in both front  and backend being set to a different format, this seems immutable) the form is set to reset after submission and a success message is shown. The user is then redirected to the Calendar page. More info to be found in Bug "Date formatting nightmare"