Refactor Calendar.jsx to dynamically render CosPlan data
This component was initially set up with static data just to test the UX of the rendering, so now that the CosPlanForm is working I needed to change the Calendar to accept the API data.

I removed all mentions of datetime as this was something I no longer wished to render, replacing it with the reference for due_date. I then put the calendar calculations for the number of events on each day into a useEffect so React could keep hold of the data.