Edit Cosplan functionality complete from Calendar
Remove setCosPlanData from the post request as it was unnecessary due to the redirect.

Add useState for tracking if the user is submitting the data or not, and once they are the form is disabled so they cannot make a post request again once the request is in flight.

Tracking if the cosplanData has an id or not, if it does, spread the previous cosplanData (so that id and due_date are not editable) and then populate the other fields with the data connected to the id.

Change success message to track if the cosplanData has an id, and display an "updated" message, or the "created" message.