Add Footer component back to solve weird loading visual issue upon mounting
I noticed that there was a split second where the page would load a giant Facebook icon before the Tailwind CSS seemed to have kicked in and mentioned it to my mentor. He said it was caused by having the footer in the index.html so I moved it out and it solved it. Now to fix the footer being at the bottom of the content rather than the bottom of the page!